+++ packages/web-forms/src/demo/demo.ts
@@ -11,7 +11,7 @@ import roboto400 from '@fontsource/roboto/400.css?inline';
 import roboto500 from '@fontsource/roboto/500.css?inline';
 // TODO/sk: Purge it - postcss-purgecss
 import primeflex from 'primeflex/primeflex.css?inline';
-import appStyles from '../assets/css/style.scss?inline';
+import appStyles from '../assets/styles/style.scss?inline';
 import router from './router';
 
 const styles = [
